Web Inspector: Elements: Unable to navigate with keyboard past `<style>` element in DOM tree https://bugs.webkit.org/show_bug.cgi?id=299445 rdar://159841729
Reviewed by Devin Rousso.

When selecting nodes in the DOM tree outline, only the suitable details sidebar panels are shown.
For a text node, like the contents of a `<style>` node, only the Node details sidebar is shown,
while Styles, Computed, Layout, Fonts, and Changes are removed.

Navigating between different nodes types triggers relayout in `WI.DOMTreeContentView`
as the details sidebars are added or removed.

The view resize handler in `WI.DOMTreeContentView.prototype.sizeDidChange()`
re-selects the selected node in order to scroll it back into the viewport if needed.
This was implemented in https://commits.webkit.org/204576@main

The DOM node is associated with both the opening and closing tags in the DOM tree outline.
But when re-selecting the DOM node, there's no context about which of the tags was previously selected.

Keyboard navigation correctly goes from the text node to the closing `<style>` tag.
But the action in the resize handler causes the the opening `<style>` to be selected. A loop is created.

There is no need to re-select the DOM node (and unnecessarily wasteful for all the sidebars),
It's enough to bring its corresponding DOM tree outline element back into view.

* Source/WebInspectorUI/UserInterface/Views/DOMTreeContentView.js:
(WI.DOMTreeContentView.prototype.sizeDidChange):
* Source/WebInspectorUI/UserInterface/Views/DOMTreeOutline.js:
(WI.DOMTreeOutline.prototype.objectForSelection):

Canonical link: https://commits.webkit.org/300471@main