Add setCDM method to AudioVideoRenderer https://bugs.webkit.org/show_bug.cgi?id=299421 rdar://161230973
Reviewed by Jer Noble.

The GPU process is always set with the entitlements that exclude this code.
As such, it's unlikely this code is still in use today.
However, in a forthcoming change, we will want the AudioVideoRenderer to be
the class attaching the CDM's session to the enqueued MediaSample as
the SourceBufferPrivateAFVObjC running in the content process is unable
to do so.
For now we copy the code from MediaPlayerMediaSourcePrivateAVFObjC

No change in existing behaviour. Covered by existing tests.

* Source/WebCore/platform/graphics/AudioVideoRenderer.h:
(WebCore::TracksRendererManager::setCDMInstance):
* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.h:
* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.mm:
(WebCore::supportsAttachContentKey):
(WebCore::shouldAddContentKeyRecipients):
(WebCore::AudioVideoRendererAVFObjC::addAudioRenderer):
(WebCore::AudioVideoRendererAVFObjC::destroyAudioRenderer):
(WebCore::AudioVideoRendererAVFObjC::ensureLayer):
(WebCore::AudioVideoRendererAVFObjC::destroyLayer):
(WebCore::AudioVideoRendererAVFObjC::setCDMInstance):

Canonical link: https://commits.webkit.org/300495@main