REGRESSION (298372@main): Occasional CheckedPtr crash when executing insertUnorderedList command https://bugs.webkit.org/show_bug.cgi?id=299492 rdar://161086162
Reviewed by Abrar Rahman Protyasha and Ryosuke Niwa.

After the changes in 298372@main, we now crash when decrementing `CheckedPtr<RenderStyle>`
underneath `EditingStyle::init`:

```
    if (node && node->computedStyle()) {
        CheckedPtr renderStyle = node->computedStyle();
        removeTextFillAndStrokeColorsIfNeeded(renderStyle.get());
        if (renderStyle->fontDescription().keywordSize()) {
            if (auto cssValue = computedStyleAtPosition.getFontSizeCSSValuePreferringKeyword())
                mutableStyle->setProperty(CSSPropertyFontSize, cssValue->cssText(CSS::defaultSerializationContext()));
        }
    }
```

This is because `getFontSizeCSSValuePreferringKeyword()` updates layout, which could potentially
destroy and recreate the node's renderer along with its `RenderStyle`. To fix this, we simply limit
the scope of the `CheckedPtr` so that it doesn't outlive the layout update.

Test: editing/execCommand/insert-unordered-list-after-DOMNodeRemoved.html

* LayoutTests/editing/execCommand/insert-unordered-list-after-DOMNodeRemoved-expected.txt: Added.
* LayoutTests/editing/execCommand/insert-unordered-list-after-DOMNodeRemoved.html: Added.

Add a test to exercise the fix; this test passes if it does not crash.

* Source/WebCore/editing/EditingStyle.cpp:
(WebCore::EditingStyle::init):

Canonical link: https://commits.webkit.org/300502@main