Validation rule for PhantomNewArrayWithButterfly is overly conservative https://bugs.webkit.org/show_bug.cgi?id=299530 rdar://161331072
Reviewed by Yusuke Suzuki.

Relax our validation rules for a PhantomNewArrayWithButterfly pointing to
a non-Phantom NewButterflyWithSize. We're seeing crashes of this on builds
but don't have any obvious repro steps. It seems like it should be fine
to have a PhantomNewArrayWithButterfly that has an allocated Butterfly
as long as nothing is actually stored into that Butterfly. This should
hopefully solve the crashes.

Canonical link: https://commits.webkit.org/300523@main