"No new tests" text has no newline after it when created via commit message hook https://bugs.webkit.org/show_bug.cgi?id=299534 rdar://161340860
Reviewed by Jonathan Bedard and Sam Sneddon.

* Tools/Scripts/hooks/prepare-commit-msg:
After 299869@main, the output of prepare-ChangeLogs now includes important newlines
(specifically, between "no new tests" and the list of changed files).

The existing code in prepare-commit-msg strips all empty lines.

I initially started by trying to teach it to preserve this line, but then realized
that all of this complexity is purely here to serve merging multiple directories'
ChangeLog's file lists together, which is no longer a concern now that we have
no physical ChangeLogs. (There is yet more complexity in both this script
and prepare-ChangeLogs itself that can be cleaned up now that this is true.) So,
instead, just delete this code, and don't have prepare-ChangeLogs add delimiters,
and maintain the formatting that it outputs.

Canonical link: https://commits.webkit.org/300525@main