[css-anchor-position-1] StyleRulePositionTry doesn't properly make a copy of itself rdar://161205381 https://bugs.webkit.org/show_bug.cgi?id=299402
Reviewed by Antti Koivisto.

WebKit implements stylesheet caching, such that two stylesheets of identical
content share the same StyleSheetContents. When JavaScript code mutates a
@position-try, CSSPositionTryDescriptors (JS object representing the rule) signals
the CSSStylesheet containing the CSSPositionTryRule that it will be mutated
(CSSPositionTryDescriptors::set* -> PropertySetCSSDescriptors::setPropertyInternal ->
PropertySetCSSDescriptors::willMutate -> CSSStyleSheet::willMutateRules).
CSSStyleSheet holds a StyleSheetContents that stores the content of the stylesheet,
and because of StyleSheetContents sharing, mutating the CSSStyleSheet requires
(1) cloning the StyleSheetContents into a separate copy, so when the stylesheet mutates,
    it doesn't affect other CSSStyleSheet sharing the StyleSheetContents
(2) Within CSSStyleSheet, each rule has a CSSOM wrapper object that attaches to
    the rule object in StyleSheetContents. After (1), the CSSOM wrapper objects have
    to be re-attached to the copied rule.

There are two bugs in StyleRulePositionTry that leads to the StyleRulePositionTry not
being copied properly:
* Cloning StyleSheetContents involves cloning all rules within it by calling ::copy()
  of the rule object. StyleRulePositionTry::copy() calls the copy constructor of
  StyleRulePositionTry, but the copy constructor is default, which doesn't deep copy
  the properties within the rule.
* (2) is done by calling CSSPositionTryRule::reattach, but it doesn't actually reattach
  itself to the new rule. So any mutation on CSSPositionTryRule ends up in the rule object
  before the clone.

If two CSSStyleSheet have identical content at the beginning, then they initially share
a StyleSheetContents. If one CSSStyleSheet content mutates, its StyleSheetContents is cloned
into a copy separate from the other CSSStyleSheet. But if the CSSStyleSheet contains a
@position-try rule, since StyleRulePositionTry doesn't clone itself properly, the new
StyleRulePositionTry and old StyleRulePositionTry ends up pointing to the same properties
object. Then, when one CSSStyleSheet mutates, the mutation is visible in the other CSSStyleSheet.

Test: imported/w3c/web-platform-tests/css/css-anchor-position/position-try-rule-caching.html

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-rule-caching-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-rule-caching.html: Added.
* Source/WebCore/css/CSSPositionTryRule.cpp:
(WebCore::StyleRulePositionTry::StyleRulePositionTry):
    - Make a copy of the StyleProperties.

(WebCore::CSSPositionTryRule::reattach):
    - Actually re-attach the CSSOM wrapper to the new rule object.

* Source/WebCore/css/CSSPositionTryRule.h:

Canonical link: https://commits.webkit.org/300553@main