REGRESSION (286540@main): google.com: Can't paste into textarea in AI mode https://bugs.webkit.org/show_bug.cgi?id=299574 rdar://158901941
Reviewed by Lily Spiniolas and Ryosuke Niwa.

The textarea element in Google search's AI mode renders a fake caret with a DOM element that runs a
periodic animation, and suppresses the native caret UI using `caret-color: transparent;`. This
allows them to paint the "caret" with different colors a few times a second.

The way this is implemented, however, causes the selection to become hosted inside a different
composited layer every time the color changes. After enabling "Selection Honors Overflow Scrolling",
this causes us to update the native selection UI frequently by calling `-selectionChanged`, due to
the fact that the caret view is constantly being reparented. Subsequently, this causes the edit menu
to immediately dismiss when tapping to try and present it.

To fix this, we simply fall back to hosting the caret selection views in `WKContentView`, in the
case where the caret is transparent. Since the selection UI is hidden anyways due to the transparent
`caret-color`, we don't actually need to host the selection in a composited view to ensure that it
visually scrolls along with subscrollable containers (and/or clipped by other composited containers
on top of it).

Test: editing/selection/ios/show-edit-menu-with-transparent-caret.html

* LayoutTests/editing/selection/ios/show-edit-menu-with-transparent-caret-expected.txt: Added.
* LayoutTests/editing/selection/ios/show-edit-menu-with-transparent-caret.html: Added.
* Source/WebKit/WebProcess/WebPage/Cocoa/WebPageCocoa.mm:
(WebKit::WebPage::getPlatformEditorStateCommon const):

Move this logic into `getPlatformEditorState`, so that we can check the caret color and other
post-layout data before finding the enclosing layer.

* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::getPlatformEditorState const):

Canonical link: https://commits.webkit.org/300554@main