4 subtests in `pointerevent_after_target_appended.html?mouse` are failing https://bugs.webkit.org/show_bug.cgi?id=299560 rdar://161362257
Reviewed by Wenson Hsieh and Abrar Rahman Protyasha.

The 4 subtests which test the events received after the child moves fail due to
a missing `pointerenter` or `mouseenter` event for the child. This is because,
before dispatching those events, we check if the target was an ancestor of the
previous element. However, the logic for this erroneously included the previous
element itself, resulting in the missing events. To resolve this, store
only the ancestors of the element under the mouse in m_ancestorsOfLastElementUnderMouse,
not the element itself.

Tested by
`imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended.html?mouse`.

* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended_interleaved.tentative_mouse-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended_interleaved.tentative_pen-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended_mouse-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_after_target_appended_pen-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_lostpointercapture_for_disconnected_node-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointer_boundary_events_after_reappending_last_over_target-expected.txt:
* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::updateMouseEventTargetNode):
* Source/WebCore/page/EventHandler.h:

Canonical link: https://commits.webkit.org/300564@main