[GTK][WPE] Use WidthIterator for simple text when Font::applyTransforms is not needed https://bugs.webkit.org/show_bug.cgi?id=299278
Reviewed by Fujii Hironori.

In 299354@main we introduced a build option to always use
ComplexTextController even for simple text code path. However, there are
cases in which we can use WidthIterator instead, for example when
optimize for speed rendering mode is used, because we know
Font::ApplyTransforms can be skipped. This patch removes that build
option, to make it possible to make the decision at runtime.

Canonical link: https://commits.webkit.org/300567@main