[iOS Firefox] Audio continues to play after closing a soundcloud.com tab https://bugs.webkit.org/show_bug.cgi?id=299545 rdar://159977164
Reviewed by Jer Noble.

When Firefox "closes" a tab in their tab overview panel, the WKWebView is hidden but not
closed. Because nothing is done to stop media playback, if audio is playing when the
user closes the tab it will continue to play. The only way to stop playback is to quit
Firefox or open a new url. Fix this by adding a linked-on-or-after Quirk to prevent
background audio playback.

* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.h:
* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.mm:
(WTF::computeSDKAlignedBehaviors):
(WTF::IOSApplication::isFirefox):
(WTF::IOSApplication::isFirefoxFocus):
* Source/WTF/wtf/spi/darwin/dyldSPI.h:
* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::initializeMediaSession):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::requirePageVisibilityToPlayAudioQuirk const):
(WebCore::Quirks::determineRelevantQuirks):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h:

Canonical link: https://commits.webkit.org/300584@main