"No new tests" text is duplicated when amending a commit https://bugs.webkit.org/show_bug.cgi?id=299557 rdar://161361290
Reviewed by Ryan Haddad.

When amending a commit, we re-run prepare-ChangeLog and merge the output.
The logic to do so was built with the expectation that prepare-ChangeLog
would only output the file listing, but this is not the case now that we
have repaired the test warning/listing.

In order to fix this, provide a separate switch to disable the test warning,
and adopt it when amending. Ideally, we'd update the test listing/warning state
based on the combined commit, but working out how to do that (since the test listing
is pretty freeform) is well outside the scope of fixing this bug.

* Tools/Scripts/hooks/prepare-commit-msg:
Avoid duplication by disabling test mentions when amending.

* Tools/Scripts/prepare-ChangeLog:
(main):
(generateNewChangeLogs):
Add a knob to disable test mentions.

Canonical link: https://commits.webkit.org/300609@main