incorrect-put-could-generate-invalid-ic-but-still-not-causing-bad-behavior-bad-transition.js's mustCrash is wrong https://bugs.webkit.org/show_bug.cgi?id=299666 rdar://161481403
Reviewed by Yijia Huang.

276881@main changed it to mustCrash which means it must crash all the
time. But this does not work because it crashes only when using IC.
If you tier up super early and avoid using IC, then it does not crash.
This patch uses runNoCJIT not to run arbitrary configurations since we
are not ensuring that all of them uses IC.

* JSTests/stress/incorrect-put-could-generate-invalid-ic-but-still-not-causing-bad-behavior-bad-transition-debug.js:
* JSTests/stress/incorrect-put-could-generate-invalid-ic-but-still-not-causing-bad-behavior-bad-transition.js:

Canonical link: https://commits.webkit.org/300665@main