AX: Hover Text does not enlarge input values https://bugs.webkit.org/show_bug.cgi?id=299130 rdar://160891414
Reviewed by Joshua Hoffman.

The reason this bug happened was because when HoverText did a hit test on an input, we returned the deepest unignored
match at the point, which was a generic div rendered inside the user-agent shadow tree generated by input elements.
There is no reason to expose this div in the accessibility tree, so ignore it, in turn causing the hit test to return
the text input, which HoverText knows how to deal with.

* LayoutTests/accessibility/text-input-hit-test-expected.txt: Added.
* LayoutTests/accessibility/text-input-hit-test.html: Added.
* LayoutTests/platform/ios/TestExpectations:
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::computeIsIgnored const):

Canonical link: https://commits.webkit.org/300691@main