First pointerdown event lost after context menu is triggered https://bugs.webkit.org/show_bug.cgi?id=232123 rdar://84787733
Reviewed by Lily Spiniolas and Abrar Rahman Protyasha.

We now explicitly do some cleanup on PointerCaptureController when
showing a contextmenu due right-click. This is required because the
events are not processed while the contextmenu is open, causing the
mousedown that opens the contextmenu to not have a matching mouseup.

The lack of a matching mouseup used to result in one later mousedown
being interpreted as part of a mouse button chord, causing
PointerCaptureController::pointerEventForMouseEvent() to produce a
pointermove instead of a pointerdown.

Canonical link: https://commits.webkit.org/300696@main