[Tools][WPE] run-mvt-tests: update default server address and test expectations https://bugs.webkit.org/show_bug.cgi?id=299821
Unreviewed gardening commit.

This updates the default remote server for the run-mvt-tests script to use
a new deployment that we have made available because the previous is gone.

Also this patch adds new expectations as well as a few minor fixes to adapt
the runner to the new version used on this new server deployment.

* Tools/Scripts/run-mvt-tests:
(parse_args):
* Tools/gtk/MVT_TestExpectations_Release.json:
* Tools/wpe/MVT_TestExpectations_Release.json:

Canonical link: https://commits.webkit.org/300710@main