[iOS] <select> element menu may dismiss early https://bugs.webkit.org/show_bug.cgi?id=299822 rdar://147261775
Reviewed by Wenson Hsieh and Abrar Rahman Protyasha.

When presenting a context menu after dismissing another one, the first menu's
dismissal animation may complete while the second menu is being presented. This
scenario is not accounted for in `WKSelectPicker`, which calls
`-[WKContentView accessoryDone]` whenever a dismissal animation is completed.
Consequently, the completion of the dismissal of the first menu can trigger the
dismissal of the second menu.

Fix by only calling `accessoryDone` after dismissal if the focused element is
still the same. If the focused element changes for any other reason, the menu
would already be dismissed.

* LayoutTests/fast/forms/ios/select-change-focus-to-another-select-expected.txt: Added.
* LayoutTests/fast/forms/ios/select-change-focus-to-another-select.html: Added.
* Source/WebKit/UIProcess/ios/forms/WKFormSelectPicker.mm:
(-[WKSelectPicker contextMenuInteraction:willEndForConfiguration:animator:]):

Canonical link: https://commits.webkit.org/300720@main