vp9HardwareDecoderAvailable() is non-functional when called in the content process. https://bugs.webkit.org/show_bug.cgi?id=299710 rdar://161534911
Reviewed by Youenn Fablet.

We re-use the VP9TestingOverrides structure to be able to store if we
the VP9 HW decoder is available. When the value of GPUProcessConnectionInfo::hasVP9HardwareDecoder
is retrieved, we set it to VP9TestingOverrides.

This allows to enable the SourceBufferParserWebM to work in the content process.

* Source/WebCore/platform/graphics/cocoa/VP9UtilitiesCocoa.h:
* Source/WebCore/platform/graphics/cocoa/VP9UtilitiesCocoa.mm:
(WebCore::VP9TestingOverrides::setVP9HardwareDecoderEnabledOverride):
(WebCore::vp9HardwareDecoderAvailable):
* Source/WebKit/WebProcess/GPU/GPUProcessConnection.cpp:
(WebKit::GPUProcessConnection::didInitialize):

Canonical link: https://commits.webkit.org/300722@main